Add files via upload
A data visualization of my analysis on a public Spotify data set, which addresses these questions:

Which artists have the most streamed songs? 
Which Artists have the most releases of tracks?
Which songs have the most streams?
What are the most common keys used?
What is the % sweet spot for danceability and energy for most EDM music?